🐛 Fix localStorage not defined

+++ pages/setting/chat-integration.vue
@@ -232,27 +232,32 @@ const { socialInfo } = await useGetSocialAccount()
 
 const cancelSocialAccount = async (id: string) => {
   try {
-    const response = await useFetch(`${import.meta.env.VITE_BASE_URL}/social-account/${id}`, {
-      method: 'delete',
-      headers: {
-        Authorization: `Bearer ${localStorage.getItem('accessToken')}`,
-      },
-    })
-    if (response.status.value === 'success') {
-      socialInfo.value.data = socialInfo.value.data.filter((social: any) => social._id !== id)
-      isSuccessDelete.value = true
-      deleteModal.value = true
-    } else {
-      console.log('call - refresh token')
-      isSuccessDelete.value = false
-      deleteModal.value = true
-      useRefreshToken()
-      cancelSocialAccount(id)
+    // Check if the code is running on the client side
+    if (process.client) {
+      const response = await useFetch(`${import.meta.env.VITE_BASE_URL}/social-account/${id}`, {
+        method: 'delete',
+        headers: {
+          Authorization: `Bearer ${localStorage.getItem('accessToken')}`,
+        },
+      })
+
+      if (response.status.value === 'success') {
+        socialInfo.value.data = socialInfo.value.data.filter((social: any) => social._id !== id)
+        isSuccessDelete.value = true
+        deleteModal.value = true
+      } else {
+        console.log('call - refresh token')
+        isSuccessDelete.value = false
+        deleteModal.value = true
+        useRefreshToken()
+        cancelSocialAccount(id)
+      }
     }
   } catch (error: any) {
     console.log(error)
   }
 }
+
 const getSocialConnectModalComponent = (socialType: string) => {
   switch (socialType) {
     case SocialType.LINE:
